apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
paho-mqtt
python-dotenv
